@@ -6,26 +6,27 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "@emailjs/browser": "^3.10.0",
-    "@material-ui/core": "4.12.3",
-    "@material-ui/icons": "4.11.2",
-    "aos": "2.3.4",
-    "prop-types": "15.7.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@emailjs/browser": "^4.4.0",
+    "@mui/icons-material": "^6.1.0",
+    "aos": "^2.3.4",
+    "prop-types": "^15.8.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-github-calendar": "^3.3.4",
-    "react-icons": "4.3.1",
-    "react-router-dom": "^6.10.0",
+    "react-icons": "^5.4.0",
+    "react-router-dom": "^6.28.0",
     "react-scripts": "^5.0.1",
-    "react-vertical-timeline-component": "3.5.2",
-    "typewriter-effect": "2.18.2"
+    "react-vertical-timeline-component": "^3.6.0",
+    "typewriter-effect": "^2.21.0"
   },
   "devDependencies": {
-    "@babel/runtime": "7.13.8",
-    "autoprefixer": "^10.4.13",
-    "postcss": "^8.4.19",
-    "tailwindcss": "^3.2.4",
-    "typescript": "4.1.3"
+    "@babel/runtime": "^7.26.0",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.15",
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "predeploy": "npm run build",
