@@ -1,6 +1,6 @@
 import React from "react";
 import "./Introduction.css";
-import ExitToAppIcon from "@material-ui/icons/ExitToApp";
+import ExitToAppIcon from "@mui/icons-material/ExitToApp";
 import { ThemeContext } from "../../Context/theme";
 
 
