@@ -6,9 +6,9 @@ import {
 } from "react-vertical-timeline-component";
 import "react-vertical-timeline-component/style.min.css";
 
-import SchoolIcon from "@material-ui/icons/School";
+import SchoolIcon from "@mui/icons-material/School";
 import { MdWork } from "react-icons/md";
-import StarRateIcon from "@material-ui/icons/StarRate";
+import StarRateIcon from "@mui/icons-material/StarRate";
 import { ThemeContext } from "../../Context/theme";
 import { Link } from "react-router-dom";
 
