@@ -1,18 +1,18 @@
 import { StrictMode } from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import { ThemeProvider } from "./Context/theme";
 import { BrowserRouter } from "react-router-dom";
 
 const rootElement = document.getElementById("root");
-ReactDOM.render(
+const root = createRoot(rootElement);
+root.render(
   <StrictMode>
     <ThemeProvider>
       <BrowserRouter>
         <App />
       </BrowserRouter>
     </ThemeProvider>
-  </StrictMode>,
-  rootElement
+  </StrictMode>
 );
